requirements.txt
alembic==1.4.0
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
awscli==1.17.10
botocore==1.14.10
Brlapi==0.6.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
Deprecated==1.2.7
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
ecdsa==0.13.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.17.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.9.2
idna==2.8
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.3.1
pycups==1.9.73
PyGithub==1.45
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.0
s3transfer==0.3.2
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
SQLAlchemy==1.3.13
system-service==0.3
systemd-python==234
typed-ast==1.3.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.2
Werkzeug==0.16.1
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
